Moved validation messages to invokable class and added interactive methods to validator class
